prometheus-operator: Move Prometheus vars in its block
This commit moves all the params that are Prometheus specific to its
special block.

This helps us organise all the Prometheus specific variables under the
`prometheus` block in the component config. Having these variables at
the root level does not delineate clearly. This change improves the UI
of the `prometheus-operator` component.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>